+++ tag-20180828/src/Models/CreateAssociatedResourceRulesRequest.php
@@ -1,0 +1,116 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Tag\V20180828\Models;
+
+use AlibabaCloud\Dara\Model;
+use AlibabaCloud\SDK\Tag\V20180828\Models\CreateAssociatedResourceRulesRequest\createRulesList;
+
+class CreateAssociatedResourceRulesRequest extends Model
+{
+    /**
+     * @var createRulesList[]
+     */
+    public $createRulesList;
+    /**
+     * @var string
+     */
+    public $ownerAccount;
+    /**
+     * @var int
+     */
+    public $ownerId;
+    /**
+     * @var string
+     */
+    public $regionId;
+    /**
+     * @var string
+     */
+    public $resourceOwnerAccount;
+    protected $_name = [
+        'createRulesList'      => 'CreateRulesList',
+        'ownerAccount'         => 'OwnerAccount',
+        'ownerId'              => 'OwnerId',
+        'regionId'             => 'RegionId',
+        'resourceOwnerAccount' => 'ResourceOwnerAccount',
+    ];
+
+    public function validate()
+    {
+        if (\is_array($this->createRulesList)) {
+            Model::validateArray($this->createRulesList);
+        }
+        parent::validate();
+    }
+
+    public function toArray($noStream = false)
+    {
+        $res = [];
+        if (null !== $this->createRulesList) {
+            if (\is_array($this->createRulesList)) {
+                $res['CreateRulesList'] = [];
+                $n1                     = 0;
+                foreach ($this->createRulesList as $item1) {
+                    $res['CreateRulesList'][$n1++] = null !== $item1 ? $item1->toArray($noStream) : $item1;
+                }
+            }
+        }
+
+        if (null !== $this->ownerAccount) {
+            $res['OwnerAccount'] = $this->ownerAccount;
+        }
+
+        if (null !== $this->ownerId) {
+            $res['OwnerId'] = $this->ownerId;
+        }
+
+        if (null !== $this->regionId) {
+            $res['RegionId'] = $this->regionId;
+        }
+
+        if (null !== $this->resourceOwnerAccount) {
+            $res['ResourceOwnerAccount'] = $this->resourceOwnerAccount;
+        }
+
+        return $res;
+    }
+
+    public function toMap($noStream = false)
+    {
+        return $this->toArray($noStream);
+    }
+
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['CreateRulesList'])) {
+            if (!empty($map['CreateRulesList'])) {
+                $model->createRulesList = [];
+                $n1                     = 0;
+                foreach ($map['CreateRulesList'] as $item1) {
+                    $model->createRulesList[$n1++] = createRulesList::fromMap($item1);
+                }
+            }
+        }
+
+        if (isset($map['OwnerAccount'])) {
+            $model->ownerAccount = $map['OwnerAccount'];
+        }
+
+        if (isset($map['OwnerId'])) {
+            $model->ownerId = $map['OwnerId'];
+        }
+
+        if (isset($map['RegionId'])) {
+            $model->regionId = $map['RegionId'];
+        }
+
+        if (isset($map['ResourceOwnerAccount'])) {
+            $model->resourceOwnerAccount = $map['ResourceOwnerAccount'];
+        }
+
+        return $model;
+    }
+}