+++ tag-20180828/src/Models/ListAssociatedResourceRulesRequest.php
@@ -1,0 +1,154 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Tag\V20180828\Models;
+
+use AlibabaCloud\Dara\Model;
+
+class ListAssociatedResourceRulesRequest extends Model
+{
+    /**
+     * @var int
+     */
+    public $maxResult;
+    /**
+     * @var string
+     */
+    public $nextToken;
+    /**
+     * @var string
+     */
+    public $ownerAccount;
+    /**
+     * @var int
+     */
+    public $ownerId;
+    /**
+     * @var string
+     */
+    public $regionId;
+    /**
+     * @var string
+     */
+    public $resourceOwnerAccount;
+    /**
+     * @var string[]
+     */
+    public $settingName;
+    /**
+     * @var string
+     */
+    public $status;
+    protected $_name = [
+        'maxResult'            => 'MaxResult',
+        'nextToken'            => 'NextToken',
+        'ownerAccount'         => 'OwnerAccount',
+        'ownerId'              => 'OwnerId',
+        'regionId'             => 'RegionId',
+        'resourceOwnerAccount' => 'ResourceOwnerAccount',
+        'settingName'          => 'SettingName',
+        'status'               => 'Status',
+    ];
+
+    public function validate()
+    {
+        if (\is_array($this->settingName)) {
+            Model::validateArray($this->settingName);
+        }
+        parent::validate();
+    }
+
+    public function toArray($noStream = false)
+    {
+        $res = [];
+        if (null !== $this->maxResult) {
+            $res['MaxResult'] = $this->maxResult;
+        }
+
+        if (null !== $this->nextToken) {
+            $res['NextToken'] = $this->nextToken;
+        }
+
+        if (null !== $this->ownerAccount) {
+            $res['OwnerAccount'] = $this->ownerAccount;
+        }
+
+        if (null !== $this->ownerId) {
+            $res['OwnerId'] = $this->ownerId;
+        }
+
+        if (null !== $this->regionId) {
+            $res['RegionId'] = $this->regionId;
+        }
+
+        if (null !== $this->resourceOwnerAccount) {
+            $res['ResourceOwnerAccount'] = $this->resourceOwnerAccount;
+        }
+
+        if (null !== $this->settingName) {
+            if (\is_array($this->settingName)) {
+                $res['SettingName'] = [];
+                $n1                 = 0;
+                foreach ($this->settingName as $item1) {
+                    $res['SettingName'][$n1++] = $item1;
+                }
+            }
+        }
+
+        if (null !== $this->status) {
+            $res['Status'] = $this->status;
+        }
+
+        return $res;
+    }
+
+    public function toMap($noStream = false)
+    {
+        return $this->toArray($noStream);
+    }
+
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['MaxResult'])) {
+            $model->maxResult = $map['MaxResult'];
+        }
+
+        if (isset($map['NextToken'])) {
+            $model->nextToken = $map['NextToken'];
+        }
+
+        if (isset($map['OwnerAccount'])) {
+            $model->ownerAccount = $map['OwnerAccount'];
+        }
+
+        if (isset($map['OwnerId'])) {
+            $model->ownerId = $map['OwnerId'];
+        }
+
+        if (isset($map['RegionId'])) {
+            $model->regionId = $map['RegionId'];
+        }
+
+        if (isset($map['ResourceOwnerAccount'])) {
+            $model->resourceOwnerAccount = $map['ResourceOwnerAccount'];
+        }
+
+        if (isset($map['SettingName'])) {
+            if (!empty($map['SettingName'])) {
+                $model->settingName = [];
+                $n1                 = 0;
+                foreach ($map['SettingName'] as $item1) {
+                    $model->settingName[$n1++] = $item1;
+                }
+            }
+        }
+
+        if (isset($map['Status'])) {
+            $model->status = $map['Status'];
+        }
+
+        return $model;
+    }
+}